+++ lib/vanagon/cli/list.rb
@@ -87,7 +87,7 @@ class Vanagon
           '--projects' => :projects,
           '--use-spaces' => :use_spaces,
         }
-        return docopt_options.map { |k, v| [translations[k], v] }.to_h
+        return docopt_options.transform_keys { |k| translations[k] }
       end
     end
   end