+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 This changelog adheres to [Keep a CHANGELOG](https://keepachangelog.com).
 
 ## [Unreleased]
+- Bump minimum ruby requirement to 2.7
+- Fix preserve param default behaving like `always` but should be `on-failure`
 
 ## [0.53.0] - 2024-09-13
 ### Added